+++ packages/backend/src/graphql/investments/position.ts
@@ -1,6 +1,6 @@
 import DataLoader from "dataloader";
 import { inArray } from "drizzle-orm";
-import type { Float, Int } from "grats";
+import type { Float } from "grats";
 
 import { currentScope } from "@/auth/session-als";
 import { db } from "@/db";
@@ -27,7 +27,7 @@ import {
 export class InvestmentReinvested {
   constructor(
     /** Units acquired via dividend reinvestments. @gqlField */
-    public readonly units: Int,
+    public readonly units: Float,
     private readonly costMinor: number,
     private readonly valueMinor: number | null,
     private readonly currency: string,
@@ -68,8 +68,8 @@ export class InvestmentPosition {
   }
 
   /** Net units held. @gqlField */
-  get units(): Int {
-    return this.s.unitsHeld as Int;
+  get units(): Float {
+    return this.s.unitsHeld as Float;
   }
 
   /** Average price paid per share currently held, excluding fees and taxes. `null` when no units are held. @gqlField */
@@ -118,7 +118,7 @@ export class InvestmentPosition {
   /** DRIP (dividend-reinvestment) activity on this position. @gqlField */
   reinvested(): InvestmentReinvested {
     return new InvestmentReinvested(
-      this.s.reinvestedUnits as Int,
+      this.s.reinvestedUnits as Float,
       this.s.reinvestedCostSum,
       reinvestedValue(this.s),
       this.currency,
